backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.2.1

# Database
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.1
python-dateutil==2.8.2

# Redis
redis[asyncio]==5.0.1
aioredis==2.0.1

# Cache & Async
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
cryptography==42.0.8

# HTTP & APIs
aiohttp==3.9.5
httpx==0.27.0

# MinIO
minio==7.2.5

# Celery
celery==5.3.6
flower==2.0.1

# Email
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.27.0

# Development
black==24.3.0
pylint==3.0.4
isort==5.13.2

# Monitoring & Logging
python-json-logger==2.0.7

# Timezone
pytz==2024.1

# Utilities
click==8.1.7
